backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
pydantic-settings
httpx
passlib[argon2]
python-jose[cryptography]
redis
rq
yfinance
pandas
